requirements.txt
torch==2.4.0
torchvision==0.19.0
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.5
opencv-python==4.11.0.86
seaborn==0.13.2
scikit-image==0.21.0
requests==2.32.3
scikit-learn==1.3.2
pyyaml==6.0.2
